third_party/blink/web_tests/platform/generic/fast/forms/date-multiple-fields/date-multiple-fields-fallback-format-expected.txt

Checks incomplete datetime formats are rejected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Valid format
PASS getUserAgentShadowTextContent(input) is "31-07-1999"
Invalid format, fallback to yyyy-MM-dd
PASS getUserAgentShadowTextContent(input) is "1999-07-31"
PASS getUserAgentShadowTextContent(input) is "1999-07-31"
PASS getUserAgentShadowTextContent(input) is "1999-07-31"
PASS successfullyParsed is true

TEST COMPLETE
